Break out dash/weather_api.py
