+++ dash/board.py
@@ -1,6 +1,5 @@
 #! /usr/bin/env python
 
-import os
 
 import httpx
 import uvicorn
@@ -12,10 +11,6 @@ app = FastAPI()
 
 
 OPEN_METEO_BASE_URL = "https://api.open-meteo.com/v1/forecast"
-WEATHER_API_BASE_URL = "http://api.weatherapi.com/v1/current.json"
-
-WEATHER_API_KEY = os.getenv("WEATHER_API_KEY")
-# assert len(WEATHER_API_KEY) == 31
 
 
 def open_meteo_fetch_forecast(zip_code: str, country_code: str = "US") -> dict:  # type: ignore
@@ -41,21 +36,6 @@ def open_meteo_fetch_forecast(zip_code: str, country_code: str = "US") -> dict: 
             strict=False,
         )
     }
-
-
-def weather_api_fetch_forecast(zip_code: str, country_code: str = "US") -> dict:  # type: ignore
-    params = {
-        "key": WEATHER_API_KEY,
-        "q": f"{zip_code},{country_code}",
-        "aqi": "yes",
-        "days": "1",  # Number of days to forecast
-        "hour": "1",  # Get hourly forecasts
-    }
-    response = httpx.get(WEATHER_API_BASE_URL, params=params)
-    print(response.status_code, type(response), 1)
-    print(response.text, 2)
-    response.raise_for_status()
-    return dict(response.json())
 
 
 @app.get("/forecast/{zip_code}")
